Remove Identify, add Ping
Unfortunately we have to compose Ping into the mox until https://github.com/libp2p/rust-libp2p/issues/2109 is fixed.
The ping interval is set to a very large value (roughly 100 years), so we don't spam nodes from the rendezvous node, but react to incoming pings.
I opted to not print the ping events because they spam the logs.