Improve log formatting
Instead of debug printing all SwarmEvents, print only the events
relevant to the rendezvous server. Use display formatting. This is what
the output looks like:
{"timestamp":"2021-07-07 15:46:10","level":"INFO","fields":{"message":"Registration expired","peer":"1AZHbYKSsasAJb4aYqHnkDQ2zzaWPAQSJ4JfZKFyFV1kAw","namespace":"xmr-btc-swap","ttl":"7200","addresses":"/ip4/12.234.222.123,/ip4/12.234.222.123"}}

Note: the addresses is displayed as a comma seperated string instead of
a JSON array because JSON logging in the tracing library does not
support arrays.